requirements.txt
pandas
matplotlib
python-dateutil
tqdm 
lxml 